requirements.txt
matplotlib
discord.py
Wavelink
pandas
Pillow
